+++ compliance/ediscovery_cases_item_review_sets_count_request_builder.go
@@ -44,9 +44,9 @@ func NewEdiscoveryCasesItemReviewSetsCountRequestBuilder(rawUrl string, requestA
 }
 // Get get the number of the resource
 // Deprecated: The ediscovery Apis are deprecated under /compliance and will stop returning data from February 01, 2023. Please use the new ediscovery Apis under /security. as of 2022-12/ediscoveryNamespace on 2022-12-05 and will be removed 2023-02-01
-// returns a *int64 when successful
+// returns a *int32 when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
-func (m *EdiscoveryCasesItemReviewSetsCountRequestBuilder) Get(ctx context.Context, requestConfiguration *EdiscoveryCasesItemReviewSetsCountRequestBuilderGetRequestConfiguration)(*int64, error) {
+func (m *EdiscoveryCasesItemReviewSetsCountRequestBuilder) Get(ctx context.Context, requestConfiguration *EdiscoveryCasesItemReviewSetsCountRequestBuilderGetRequestConfiguration)(*int32, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
         return nil, err
@@ -54,14 +54,14 @@ func (m *EdiscoveryCasesItemReviewSetsCountRequestBuilder) Get(ctx context.Conte
     errorMapping := i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.ErrorMappings {
         "XXX": i20a3050780ee0b0cde0a884a4f35429a20d60067e3bcda382ec5400079147459.CreateODataErrorFromDiscriminatorValue,
     }
-    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int64", errorMapping)
+    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int32", errorMapping)
     if err != nil {
         return nil, err
     }
     if res == nil {
         return nil, nil
     }
-    return res.(*int64), nil
+    return res.(*int32), nil
 }
 // ToGetRequestInformation get the number of the resource
 // Deprecated: The ediscovery Apis are deprecated under /compliance and will stop returning data from February 01, 2023. Please use the new ediscovery Apis under /security. as of 2022-12/ediscoveryNamespace on 2022-12-05 and will be removed 2023-02-01
